### 2025-02-26 #7848 @jonsampedro
Thank you, @kthyng!

There don’t seem to be any DOI or paper-related issues, but please let me know if you need any additional information.

For reviewers, I suggest the following:

- Paul Taconet
- Nan-Hung Hsieh
- slwu89
- Ciara Donegan
- nistara
- Longhao Wang
- meenakshi-kushwaha

Thanks again!

Best,
Jon
