### 2025-03-21 #7848 @kthyng
@jonsampedro Sorry I haven't had a chance to get to processing your submission fully yet, but I will do so soon. In the meantime you could shorten your paper to be closer to 1000 words.
