### 2025-03-31 5ac6a44
shorten paper for JOSS requirements
