+++ paper/paper.bib
@@ -1,30 +1,9 @@
-@report{tracking_2024,
-	title = {Tracking SDG 7: The Energy Progress Report},
-	author = {IEA, World Bank, WHO, , UNSD},
-	year = {2024},
-        publisher={World Bank Washington, DC}
-}
-
-@article{balakrishnan2013state,
-  title={State and national household concentrations of PM 2.5 from solid cookfuel use: results from measurements and modeling in India for estimation of the global burden of disease},
-  author={Balakrishnan, Kalpana and Ghosh, Santu and Ganguli, Bhaswati and Sambandam, Sankar and Bruce, Nigel and Barnes, Douglas F and Smith, Kirk R},
-  journal={Environmental Health},
-  volume={12},
-  pages={1--14},
-  year={2013},
-  publisher={Springer},
-  DOI = {10.1186/1476-069X-12-77}
-}
-
-@article{bennitt2021estimating,
-  title={Estimating disease burden attributable to household air pollution: new methods within the Global Burden of Disease Study},
-  author={Bennitt, FB and Wozniak, SS and Causey, Kate and Burkart, Katrin and Brauer, Michael},
-  journal={The Lancet Global Health},
-  volume={9},
-  pages={S18},
-  year={2021},
-  publisher={Elsevier},
-  DOI = {10.1016/S2214-109X(21)00126-1}
+@article{bennitt2025global,
+  title={Global, regional, and national burden of household air pollution, 1990--2021: a systematic analysis for the Global Burden of Disease Study 2021},
+  author={Bennitt, Fiona B and Wozniak, Sarah and Causey, Kate and Spearman, Sandra and Okereke, Chukwuma and Garcia, Vanessa and Hashmeh, Nadim and Ashbaugh, Charlie and Abdelkader, Atef and Abdoun, Meriem and others},
+  journal={The Lancet},
+  year={2025},
+  publisher={Elsevier}
 }
 
 @article{neidell2023air,
