+++ paper/paper.md
@@ -72,28 +72,20 @@ map = T)
 
 # Statement of need
 
-Recent data indicates that 2.1 billion people worldwide lack access to clean, modern cooking fuels and technologies [@tracking_2024]. 
-As a result, they rely on open fires or inefficient stoves fueled by kerosene, traditional biomass (such as fuelwood or animal dung), and coal. 
-These conventional energy sources release a range of harmful pollutants, including fine particulate matter (PM2.5), nitrogen oxides (NOx), and volatile organic compounds (VOCs). 
-These direct emissions, coupled with factors such as household size, ventilation, and building materials [@balakrishnan2013state], are the primary contributors to the formation of HAP.
-
 According to the World Health Organization’s (WHO) [Global Health Observatory]( https://www.who.int/data/gho/data/indicators/indicator-details/GHO/household-air-pollution-attributable-deaths), 
 HAP is one of the major risk factors for human health, particularly in the global South, causing around 3.2 million deaths per year in 2020, including over 237,000 deaths of children under the age of 5. 
-HAP affects human health from pre-conception to old age, leading to noncommunicable diseases including cataract, chronic obstructive pulmonary disease, ischaemic heart disease, lower respiratory infections, lung cancer, neonatal disorders, stroke, and type 2 diabetes [@bennitt2021estimating]. 
-In addition to its direct health impacts, which disproportionately affect infants and the elderly, HAP also has broader socioeconomic consequences.
-For instance, recent studies demonstrate that exposure to high pollution levels can significantly reduce work productivity [@neidell2023air]. 
-Likewise, women bear the greatest health burden from polluting fuels, as they often handle household chores, further exacerbating gender inequalities [@krishnapriya2021improved]. 
+HAP affects human health from pre-conception to old age, leading to noncommunicable diseases including cataract, chronic obstructive pulmonary disease, ischaemic heart disease, lower respiratory infections, lung cancer, neonatal disorders, stroke, and type 2 diabetes [@bennitt2025global]. 
+In addition to its direct health impacts, disproportionately affecting infants and the elderly, household air pollution (HAP) also has broader socioeconomic consequences, 
+such as reduced work productivity [@neidell2023air] and the exacerbation of gender inequalities [@krishnapriya2021improved].
 
-Given the magnitude of these effects, various methodologies have been developed to estimate HAP and its associated health implications.
-A recent study employed Bayesian hierarchical models to estimate annual average personal exposure and indoor concentration levels [@mohajeri2023urban]. 
-Similarly, the BAR-HAP tool, developed by the WHO, provides a framework to quantify the benefits of reducing HAP and compare them against the costs of intervention [@das2021benefits]. 
+Given the magnitude of these effects, various methodologies have been developed to estimate HAP and its associated health implications [@mohajeri2023urban; @das2021benefits]. 
 However, HAP and its subsequent impacts on human health are usually ignored in global scenario analysis. 
 Alternative socioeconomic and climate scenarios, such as high or low GDP growth pathways and the adoption of deep decarbonization strategies, will significantly influence the transition to cleaner energy fuels and technologies.
 These pathways can enhance access to less harmful energy sources for lower-income households, thereby reducing exposure to HAP and mitigating its adverse impacts on human health.
 The incorporation of these dynamics into integrated assessment models, such as the Global Change Analysis Model (GCAM), offers valuable insights for evaluating alternative future scenarios.
 
 GCAM is a multisector integrated assessment model that quantifies human and Earth-system dynamics by examining the interconnections between the economy, energy, water, climate, and agriculture, forestry, and land use (AFOLU) systems. 
[…]
