+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   fit_model: fit_model.html
   ResidEm_grp: ResidEm_grp.html
   run_rhap: run_rhap.html
-last_built: 2025-02-25T13:12Z
+last_built: 2025-03-31T07:22Z
 urls:
   reference: https://bc3lc.github.io/rhap/reference
   article: https://bc3lc.github.io/rhap/articles
