+++ paper/paper.md
@@ -89,8 +89,6 @@ Detailed documentation is available [online](https://github.com/JGCRI/gcam-doc).
 Focusing on emissions, GCAM estimates a wide range of greenhouse gases and air pollutants for each future scenario, categorized by sector, region, and time period up to 2100. 
 Nevertheless, while GCAM estimates direct air pollutant emissions from the residential sector, it does not account for HAP levels or their associated health impacts. 
 The `rhap` package bridges this gap by quantifying HAP impacts for each GCAM scenario. 
-Using an econometric model detailed in the provided [documentation]( https://bc3lc.github.io/rhap/articles/fit_model.html), `rhap` extracts emissions and other relevant data, such as region- and period-specific floorspace and per capita GDP, from a given GCAM output database. 
-It then produces detailed tables, figures, and maps illustrating HAP impacts by country and time period for the entire world. 
 Given the well-documented importance of HAP and its health impacts, a tool that automatically estimates these effects within the GCAM modelling framework represents a valuable advancement for the integrated assessment modelling community, 
 and it addresses a critical gap in global scenario analysis.
 
@@ -106,10 +104,8 @@ Detailed information about the model’s formulation, underlying assumptions, an
 The `calc_hap_impacts` function offers flexibility in generating three distinct health impact metrics: premature mortalities, Years of Life Lost (YLLs), and Disability-Adjusted Life Years (DALYs). 
 Users can specify the desired metric by setting the `HIA_var` parameter, with premature mortalities selected by default. 
 The package also offers an optional feature (`by_gr = TRUE`) to estimate health impacts by income decile within each region. While country-level calculations ensure consistency, this feature helps explore intra-regional disparities.
-For further customization, users can adjust the `normalized` parameter to produce results either as absolute values or normalized metrics (e.g., per 100,000 inhabitants), catering to different analytical needs. 
 In terms of outputs, activating the saveOutput parameter enables the function to save results as a Comma-Separated Values (CSV) file in the `output` sub-directory. 
 The function can also generate damage maps and animations by enabling the `map` and `anim` parameters, respectively, based on the rmap package [@khan2022rmap].
-Figure 2 illustrates a representative example of the outputs generated by the `calc_hap_impacts` function.
 
 ![Premature deaths per 100.000 inhabitants attributable to household air pollution in 2050](map_base_2050.png)
 
