### 2025-03-31 #7848 @jonsampedro
@kthyng I have reduced the paper to 1150 words. Just let me know if you need anything else from my side.

Thank you!
