### 2025-04-22 #7848 @goldingn
Hi @ibarraespinosa would you be interested in reviewing this paper on household air pollution for JOSS?
