### 2025-04-23 #8091 @editorialbot
[REVIEW]: rhap: A tool to analyze health impacts attributable to household air pollution
**Submitting author:** @jonsampedro (<a href="http://orcid.org/0000-0002-2277-1530">Jon Sampedro</a>)
**Repository:** https://github.com/bc3LC/rhap
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
