### 2025-05-07 #8091 @goldingn
Hi @ibarraespinosa, @slwu89 to get started with your reviews, please create your checklist typing:
`@editorialbot generate my checklist`

For any specific issues related to the software, you can open an issue at  https://github.com/bc3LC/rhap and mention this review thread so the two are connected.
