### 2025-06-03 bc3LC/rhap#1 @slwu89
JOSS review comments
Hello, I am posting review commends related to https://github.com/openjournals/joss-reviews/issues/8091#issuecomment-2907822288 here:

- [x] General checks: Contribution and authorship: The first and (mostly) second authors of the JOSS paper have contributed to the repository based on the contribution graph, but it is not clear what the contribution of the remaining authors is.
- [x] Functionality: Functionality: The following check box is similar to this one. While I can install the package, the examples in the documentation do not show how to use it with other packages to reproduce the results from examples (e.g. the map in https://bc3lc.github.io/rhap/articles/run_rhap.html). The documentation should be refactored such that a new user can run the entire example based on code on the example page.
- [x] Documentation: Example usage: see above
- [x] Documentation: Community guidelines: There should be a section with those guidelines for community interaction (how to report bugs, how to contribute, how to ask questions, etc) as outlined in the review thread tasks
- [x] Software paper: State of the field: The software paper does not say if there are other software libraries which fulfill a similar role, or if not, that this is entirely novel.
