### 2025-06-13 7bf49f5
add example prj file and information to provide a full step-by-step example
