+++ pkgdown.yml
@@ -1,11 +1,11 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   fit_model: fit_model.html
   ResidEm_grp: ResidEm_grp.html
   run_rhap: run_rhap.html
-last_built: 2025-03-31T07:28Z
+last_built: 2025-06-13T11:48Z
 urls:
   reference: https://bc3lc.github.io/rhap/reference
   article: https://bc3lc.github.io/rhap/articles
