+++ CONTRIBUTING.md
@@ -1,18 +1,17 @@
 ## How to contribute
-___
 
-__I found a bug!__
+### I found a bug!
 * Nice sleuthing!
-* Check if the bug has already been reported by searching the existing GitHub [Issues]. If you find a match, consider adding additional details to the existing ticket.
-* Open a [new Issue], being sure to include a clear title and description along with as much detail as possible; code samples or log messages demonstrating the bug are quite helpful.
+* Check if the bug has already been reported by searching the existing GitHub [Issues](https://github.com/bc3LC/rhap/issues). If you find a match, consider adding additional details to the existing ticket.
+* Open a [new Issue](https://github.com/bc3LC/rhap/issues/new), being sure to include a clear title and description along with as much detail as possible; code samples or log messages demonstrating the bug are quite helpful.
 
-__I fixed a bug!__
+### I fixed a bug!
 * First of all, thanks!
-* Open a [new Pull Request] with the fix. Ensure the description clearly outlines the bug and the solution. Include the Issue number if applicable.
+* Open a [new Pull Request](https://github.com/bc3LC/rhap/pulls) with the fix. Ensure the description clearly outlines the bug and the solution. Include the Issue number if applicable.
 
-__I created a new feature!__
+### I created a new feature!
 * You're the best!
-* Consider opening a [new Issue] to describe use cases for the new feature. This will offer a platform for discussion and critique.
-* Then, open a [new Pull Request] with clear documentation of the methodology. Be sure to include new unit tests if appropriate.
+* Consider opening a [new Issue](https://github.com/bc3LC/rhap/issues/new)to describe use cases for the new feature. This will offer a platform for discussion and critique.
+* Then, open a [new Pull Request](https://github.com/bc3LC/rhap/pulls) with clear documentation of the methodology. Be sure to include new unit tests if appropriate.
 
-The [IM3] team truly appreciates and encourages community involvement. We're all in this together! 
+The [BC3](https://www.bc3research.org/) team truly appreciates and encourages community involvement. We're all in this together! 
