### 2025-06-13 f74bad3
delete unneccsary libs
