+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   fit_model: fit_model.html
   ResidEm_grp: ResidEm_grp.html
   run_rhap: run_rhap.html
-last_built: 2025-06-13T11:55Z
+last_built: 2025-06-13T12:00Z
 urls:
   reference: https://bc3lc.github.io/rhap/reference
   article: https://bc3lc.github.io/rhap/articles
