+++ paper/paper.bib
@@ -60,7 +60,15 @@
   DOI = {10.21105/joss.04015}
 }
 
-
+@article{sampedro2022rfasst,
+  title={rfasst: An R tool to estimate air pollution impacts on health and agriculture},
+  author={Sampedro, Jon and Khan, Zarrar and Vernon, Chris R and Smith, Steven J and Waldhoff, Stephanie and Van Dingenen, Rita},
+  journal={Journal of Open Source Software},
+  volume={7},
+  number={69},
+  pages={3820},
+  year={2022}
+}
 
 
 
