### 2025-06-13 bc3LC/rhap#1 @jonsampedro
@slwu89 Thank you for your comments. We have changed the documentation website and the main paper following your comments and recommendations. Below you can find the detailed response to each of the comments raised.

**Reponse to comment 1**

While the technical development of the package was primarily carried out by the first two co-authors, Jon Sampedro and Clàudia Rodés-Bachs, the creation of this tool has been a collaborative effort, with valuable contributions from a diverse team.

Co-authors Manuel Tomás and Ismael Urrutia played a key role in the development of the econometric model underlying the package's core calculations, particularly in data collection, model validation, and robustness testing. Steven Smith and Dirk-Jan Van de Ven provided essential support in interpreting the GCAM results and transforming these outputs into a format suitable for integration into the econometric model. They also contributed significantly to the conceptualization of the overall package.

We believe these contributions are in line with the JOSS authorship guidelines, which state that “_active project direction and other forms of non-code contributions are (considered sufficient for co-authorship)_”.

**Reponse to comment 2 and 3** (addressed jointly)

Thank you for raising this points. Following your recommendation, we have added a complete step-by-step example to the documentation, demonstrating how to use the tool to estimate health impacts from a GCAM scenario. You can find the example here: https://bc3lc.github.io/rhap/articles/run_rhap.html#step-by-step-example

To make it easier for new users to run the example, we now provide a _.dat_ file based on a GCAM 7.1 baseline scenario, as suggested in the review.

**Reponse to comment 4**

Thank you. We have added a new “Contributing” file with the community guidelines: https://github.com/bc3LC/rhap/blob/main/CONTRIBUTING.md

**Reponse to comment 5**

We have revised the main paper following this recommendation in order to clarify the novelty and the contribution of this package:

_“Given the magnitude of these effects, various methodologies have been developed to estimate HAP and its associated health implications. However, HAP and its subsequent impacts on human health are often overlooked in global scenario analysis. Alternative socioeconomic and climate pathways, such as those involving high or low GDP growth and the adoption of deep decarbonization strategies, can significantly influence the transition to cleaner energy fuels and technologies. These transitions may improve access to cleaner energy sources for lower-income households, thereby reducing exposure to HAP and mitigating its health impacts.  Despite this, the dominant tools used in global scenario analysis, namely, integrated assessment models (IAMs), typically estimate future emissions of air pollutants but do not report on air pollution levels or their health consequences. While some external tools can be linked to IAMs to assess ambient air pollution and its impacts, there is currently no such tool available for HAP and its associated effects. Incorporating HAP dynamics into IAMs, such as the Global Change Analysis Model (GCAM), represents a meaningful contribution to the field. It provides important insights for evaluating alternative future scenarios with greater attention to equity, health, and household-level energy transitions”_
