### 2025-06-16 67a3388
add plm pkg dependency
