+++ .github/workflows/build.yaml
@@ -40,7 +40,7 @@ jobs:
         run: |
           install.packages(c("remotes","rcmdcheck","stringi"), repos = "https://cloud.r-project.org")
           install.packages(c("evaluate"), repos = "https://cloud.r-project.org")
-          install.packages(c("knitr","rmarkdown","htmlwidgets"), repos = "https://cloud.r-project.org")
+          install.packages(c("knitr","rmarkdown","htmlwidgets", "plm"), repos = "https://cloud.r-project.org")
           remotes::install_deps(dependencies = TRUE)
         shell: Rscript {0}
 
