+++ vignettes/run_rhap.Rmd
@@ -66,7 +66,7 @@ library(rhap)
 ```r
 # Set your project and scenario names
 my_prj_name <- "test_rhap.dat"  # Name of the .dat project file (avoid spaces)
-my_scen_name <- "Ref_7p1"       # Name of the GCAM scenario to process (or vector of scenarios)
+my_scen_name <- "Reference"       # Name of the GCAM scenario to process (or vector of scenarios)
 
 # Run the RHAP model
 hap_damages <- calc_hap_impacts(
