### 2025-06-16 c386934
clean INf values
