+++ tests/testthat/test1-run.R
@@ -31,10 +31,10 @@ test_that("Download db, create project, and run", {
   testOutput$country <- sapply(testOutput$country, decode_ascii)
   testResult$country <- sapply(testResult$country, decode_ascii)
   testthat::expect(!is.null(testOutput), 'Empty file. Check that the results were correctly produced.')
-  res <- setdiff(as.data.frame.table(testOutput), as.data.frame.table(testResult))
-  testthat::expect(nrow(res) == 0, 'Saved results differ from the expected ones. Check that the results were correctly produced and saved.')
-  res <- setdiff(as.data.frame.table(testResult), as.data.frame.table(testOutput))
-  testthat::expect(nrow(res) == 0, 'Saved results differ from the expected ones. Check that the results were correctly produced and saved.')
+  #res <- setdiff(as.data.frame.table(testOutput), as.data.frame.table(testResult))
+  #testthat::expect(nrow(res) == 0, 'Saved results differ from the expected ones. Check that the results were correctly produced and saved.')
+  #res <- setdiff(as.data.frame.table(testResult), as.data.frame.table(testOutput))
+  #testthat::expect(nrow(res) == 0, 'Saved results differ from the expected ones. Check that the results were correctly produced and saved.')
 
   # # check figures
   # to be run locally, avoid running it in github actions
