### 2025-07-06 bc3LC/rhap#2 @ibarraespinosa
review joss
[10.21105.joss.08091_revsie1.pdf](https://github.com/user-attachments/files/21091731/10.21105.joss.08091_revsie1.pdf)

I started the review of the submissions here https://github.com/openjournals/joss-reviews/issues/8091

This is the manuscript.

i will start checking the code
