### 2025-07-06 bc3LC/rhap#2 @ibarraespinosa
reamde.. how to guides?
