+++ README.md
@@ -5,7 +5,7 @@
 [![docs](https://github.com/bc3LC/rhap/actions/workflows/docs.yaml/badge.svg)](https://github.com/bc3LC/rhap/actions/workflows/docs.yaml)
 [![pages-build-deployment](https://github.com/bc3LC/rhap/actions/workflows/pages/pages-build-deployment/badge.svg)](https://github.com/bc3LC/rhap/actions/workflows/pages/pages-build-deployment)
 [![test_coverage](https://github.com/bc3LC/rhap/actions/workflows/test_coverage.yaml/badge.svg)](https://github.com/bc3LC/rhap/actions/workflows/test_coverage.yaml)
-[![codecov](https://codecov.io/gh/bc3LC/rhap/graph/badge.svg?token=rC8eIjNwcN)](https://codecov.io/gh/bc3LC/rhap)
+[![codecov](https://codecov.io/gh/bc3LC/rhap/branch/main/graph/badge.svg?token=rC8eIjNwcN)](https://codecov.io/gh/bc3LC/rhap)
 [![DOI](https://zenodo.org/badge/721541306.svg)](https://doi.org/10.5281/zenodo.14423225)
 
 
