### 2025-07-08 bc3LC/rhap#2 @jonsampedro
Hi @ibarraespinosa. Thank you for your comments. Some answers below:

- The "_How to_" link in the README file has been now corrected in [this commit](https://github.com/bc3LC/rhap/commit/503302e117780c2b81324db1a7a3bf5403dbf518)

- The [codecov](https://app.codecov.io/gh/bc3LC/rhap) has also been corrected.

- The [.dat file](https://github.com/bc3LC/rhap/blob/main/test_rhap.dat) to run the [Step-by-step example](https://bc3lc.github.io/rhap/articles/run_rhap.html) is provided in the repository. This can be used to run all the functions in the package.

Just let me know if you have any issue you when running the package functions. Thank you again for your time and feedback,
