### 2025-07-11 bc3LC/rhap#1 @slwu89
Hi @jonsampedro, thank you, I believe comments 1,4,5 are addressed.

For comment 2, I see that in the root directory of the repo there is the file "test_rhap.dat" which can be used in the step by step example (https://bc3lc.github.io/rhap/articles/run_rhap.html#step-by-step-example). However, it is not clear how to access it, especially for a user who (in the future) may download from CRAN and does not see the repo structure. I suggest storing it either in the data/ subdirectory so that the user may load it directly in R `data(test_rhap)` and then proceed with the example, or store it perhaps in the inst/extdata/ director where it will be available for loading into the user's workspace. For example, `system.file("extdata","iso_region.csv", package="rhap")` would get the file's location on the user machine if it were stored in extdata. I am not sure what command should be used to load in files of .dat extension, but of course the appropriate function to load it into R's workspace should also be used so that it can be run as smoothly as possible.
