### 2025-07-12 bc3LC/rhap#2 @ibarraespinosa
Installation guide is still confusing. For instance, cloning the repo is optional, but required to run the examples with the .dat file. Guides and examples should provide link and steps to download the .dat file
