### 2025-07-14 5a12a02
changes to dat file and installation guide
