+++ README.md
@@ -62,29 +62,59 @@ TBA
 
 [Back to Contents](#Contents)
 
-1. Download and install:
-    - R (https://www.r-project.org/)
-    - R studio (https://www.rstudio.com/)  
-    - (For cloning the repo) Git (https://git-scm.com/downloads) 
+The package can be directly installed following these steps:
+
+1. Install Required Software: 
+- **R**: [Download R](https://www.r-project.org/)
+- **RStudio** (recommended IDE): [Download RStudio](https://www.rstudio.com/)
+- **Git** (required to clone the repository): [Download Git](https://git-scm.com/downloads)
     
     
-2. Open R studio:
+2. Install the Package via `devtools`
+
+Open **RStudio** and run the following commands in the Console:
 
 ```r
-install.packages("devtools")
+install.packages("devtools")  # if not already installed
 devtools::install_github("bc3LC/rhap")
 ```
 
-(Optional) 
+3. Clone the Repository (Optional but Recommended)
 
-To clone the repository to the local machine: Git bash in the working directory (right click "Git Bash Here") -> In the Git console type:  
+Cloning the repository is recommended to access the full package structure and all input files. 
+
+To clone the repository: Open Git Bash in your desired working directory (Right-click → Git Bash Here).
+
+Run the following command: 
 
 ```r
 git clone https://github.com/bc3LC/rhap.git
 ```
 
-Then, open the Rproject (rfasst.Rproj): In the Rstudio menu, click "Build -> Install and restart" (Ctrl+Shift+B)
-  
+This will create a local copy of the repository.
+
+4. Load the Project in RStudio
+
+After cloning the repository:
+
+- Navigate to the cloned folder and open the rhap.Rproj file in RStudio.
+- In RStudio, go to the menu: Build → Install and Restart, or use the keyboard shortcut: Ctrl + Shift + B.
+
+Alternatively, you can load the package directly by running:
+
+```r
+devtools::load_all()
+```
+
+5. Test a Baseline Example (Optional)
+
+The package includes a .dat file that allows you to run a test case for a GCAM baseline scenario: https://bc3lc.github.io/rhap/articles/run_rhap.html#step-by-step-example 
+
+Note that this example can be run without cloning the full repository by loading the included file:
+
+```r
+test_rhap <- rhap::test_rhap
+``` 
 
 <!-- ------------------------>
 <!-- ------------------------>
