### 2025-07-14 bc3LC/rhap#1 @jonsampedro
Hi @slwu89, thank you for your comments.

As suggested, the `test_rhap` file is now stored in the `inst/extdata/` directory and included in the `data` folder, so it can be directly loaded. This has been clarified in the corresponding [vignette](https://bc3lc.github.io/rhap/articles/run_rhap.html#step-by-step-example). In addition, the [installation guide](https://github.com/bc3LC/rhap?tab=readme-ov-file#InstallGuide) has been completely restructured for improved clarity.
