### 2025-07-14 bc3LC/rhap#2 @jonsampedro
Hi @ibarraespinosa ,

The [installation guide](https://github.com/bc3LC/rhap?tab=readme-ov-file#InstallGuide) has now been completely re-structured to improve clarity.

Also, the `test_rhap` file is now stored in the `inst/extdata/` directory and included in the `data` folder, so it can be directly loaded. Now there is no need to clone the package to access the test file, as also clarified in the [vignette](https://bc3lc.github.io/rhap/articles/run_rhap.html#step-by-step-example)
