### 2025-07-27 bc3LC/rhap#2 @goldingn
Hi @ibarraespinosa are you happy with the edits to the installation guide and data?
