### 2025-08-01 bc3LC/rhap#2 @ibarraespinosa
I'm not seeing the plots, what I'm missing?

```
library(rhap)
# Set your project and scenario names
my_prj_name  <- rhap::test_rhap  # Name of the .dat project file (avoid spaces). Note that rhap::test_rhap loads the example project file included in the package

my_scen_name <- "Reference"       # Name of the GCAM scenario to process (or vector of scenarios)

# Run the RHAP model
hap_damages <- calc_hap_impacts(
  prj_name     = my_prj_name,
  scen_name    = my_scen_name,
  final_db_year = 2050,
  saveOutput   = TRUE,
  map          = FALSE,
  anim         = TRUE,
  HIA_var      = "deaths",
  normalized   = TRUE,
  by_gr        = FALSE
)

# Preview the results
head(hap_damages)

```

```
sessionInfo()
R version 4.4.1 (2024-06-14)
Platform: x86_64-pc-linux-gnu
Running under: Ubuntu 24.10

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.12.0
LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.12.0

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C
 [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
 [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

time zone: America/Denver
tzcode source: system (glibc)

attached base packages:

other attached packages:
[1] rhap_0.1.0

 [1] gcamdata_5.1      dplyr_1.1.4       compiler_4.4.1    tidyselect_1.2.1
 [5] Rcpp_1.1.0        collapse_2.1.2    xml2_1.3.8        stringr_1.5.1
 [9] parallel_4.4.1    assertthat_0.2.1  tidyr_1.3.1       lattice_0.22-6
[13] readr_2.1.5       R6_2.6.1          maxLik_1.5-2.1    generics_0.1.4
[17] lmtest_0.9-40     Formula_1.2-5     rbibutils_2.3     MASS_7.3-61
[21] bdsmatrix_1.3-7   tibble_3.3.0      lubridate_1.9.4   pillar_1.11.0
[25] tzdb_0.5.0        rlang_1.1.6       utf8_1.2.6        stringi_1.8.7
[29] rgcam_1.2.0       timechange_0.3.0  cli_3.6.5         withr_3.0.2
[33] magrittr_2.0.3    Rdpack_2.6.4      digest_0.6.37     grid_4.4.1
[37] rstudioapi_0.17.1 plm_2.6-6         hms_1.1.3         sandwich_3.1-1
[41] nlme_3.1-166      lifecycle_1.0.4   miscTools_0.6-28  vctrs_0.6.5
[45] glue_1.8.0        data.table_1.17.8 zoo_1.8-14        purrr_1.1.0
[49] httr_1.4.7        tools_4.4.1       pkgconfig_2.0.3
>
```
