### 2025-08-01 bc3LC/rhap#2 @ibarraespinosa
Well, now I can run the example, that is a step forward
