### 2025-08-08 bc3LC/rhap#2 @ibarraespinosa
https://github.com/bc3LC/rhap/blob/main/R/create_panel.R#L74 Why not using

```
system.file("extdata", package = "rhap")
```
