### 2025-08-18 2133dc7
update README: remotes pkg
