+++ inst/extdata/saveDataFiles.R
@@ -12,78 +12,78 @@ decode_ascii <- function(text) {
 #=========================================================
 # Pollutants to be included in the model
 panel_pollutants <- c("BC", "OC", "PrimPM25", "NOx", "VOC", "NMVOC")
-usethis::use_data(panel_pollutants, overwrite = T)
+usethis::use_data(panel_pollutants, overwrite = TRUE)
 
 # Panel data
 panel_data <- create_panel()
-usethis::use_data(panel_data, overwrite = T)
+usethis::use_data(panel_data, overwrite = TRUE)
 
 #Raw ssp data
 raw.ssp.data <- read.csv("./inst/extdata/socioeconomic/SSP_database_2024.csv",
                         skip = 8, fileEncoding = "UTF-8-BOM")
 raw.ssp.data$Region <- sapply(raw.ssp.data$Region, decode_ascii)
-usethis::use_data(raw.ssp.data, overwrite = T)
+usethis::use_data(raw.ssp.data, overwrite = TRUE)
 
 # Country-level GDP by SSP scenario
 gdp_ctry.SSP1 <- get_gdp_ctry(ssp = 'SSP1')
-usethis::use_data(gdp_ctry.SSP1, overwrite = T)
+usethis::use_data(gdp_ctry.SSP1, overwrite = TRUE)
 gdp_ctry.SSP2 <- get_gdp_ctry(ssp = 'SSP2')
-usethis::use_data(gdp_ctry.SSP2, overwrite = T)
+usethis::use_data(gdp_ctry.SSP2, overwrite = TRUE)
 gdp_ctry.SSP3 <- get_gdp_ctry(ssp = 'SSP3')
-usethis::use_data(gdp_ctry.SSP3, overwrite = T)
+usethis::use_data(gdp_ctry.SSP3, overwrite = TRUE)
 gdp_ctry.SSP4 <- get_gdp_ctry(ssp = 'SSP4')
-usethis::use_data(gdp_ctry.SSP4, overwrite = T)
+usethis::use_data(gdp_ctry.SSP4, overwrite = TRUE)
 gdp_ctry.SSP5 <- get_gdp_ctry(ssp = 'SSP5')
-usethis::use_data(gdp_ctry.SSP5, overwrite = T)
+usethis::use_data(gdp_ctry.SSP5, overwrite = TRUE)
 
 # Country-level population by SSP scenario
 pop_ctry.SSP1 <- get_pop_ctry(ssp = 'SSP1')
-usethis::use_data(pop_ctry.SSP1, overwrite = T)
+usethis::use_data(pop_ctry.SSP1, overwrite = TRUE)
 pop_ctry.SSP2 <- get_pop_ctry(ssp = 'SSP2')
-usethis::use_data(pop_ctry.SSP2, overwrite = T)
+usethis::use_data(pop_ctry.SSP2, overwrite = TRUE)
 pop_ctry.SSP3 <- get_pop_ctry(ssp = 'SSP3')
-usethis::use_data(pop_ctry.SSP3, overwrite = T)
+usethis::use_data(pop_ctry.SSP3, overwrite = TRUE)
 pop_ctry.SSP4 <- get_pop_ctry(ssp = 'SSP4')
-usethis::use_data(pop_ctry.SSP4, overwrite = T)
+usethis::use_data(pop_ctry.SSP4, overwrite = TRUE)
 pop_ctry.SSP5 <- get_pop_ctry(ssp = 'SSP5')
-usethis::use_data(pop_ctry.SSP5, overwrite = T)
+usethis::use_data(pop_ctry.SSP5, overwrite = TRUE)
 
 # Adjust country names due to differences in datasets:
[…]
