+++ R/get_pop_ctry.R
@@ -1,6 +1,6 @@
 #' get_pop_ctry
 #'
-#'@description
+#' @description
 #' Ancillary function to get country-level GDPpc from SSP database
 #'
 #' SSP database version 3.0.1 updated in 2024
@@ -13,7 +13,6 @@
 #' @return Per capita GDP by country
 
 get_pop_ctry <- function(ssp) {
-
   iso <- country_name <- year <- pop <- continent <- dev <- variable <- value <-
     Model <- Scenario <- Region <- Variable <- Unit <- region <- scenario <- . <- NULL
 
@@ -29,12 +28,12 @@ get_pop_ctry <- function(ssp) {
   conv_M_people <- 1E6
 
   # Ancillary Functions
-  `%!in%` = Negate(`%in%`)
+  `%!in%` <- Negate(`%in%`)
 
   # First, we read in the population data.
   max_base_year <- rhap::raw.ssp.data %>%
     tidyr::gather(year, value, -Model, -Scenario, -Region, -Variable, -Unit) %>%
-    dplyr::mutate(year=gsub("X", "", year)) %>%
+    dplyr::mutate(year = gsub("X", "", year)) %>%
     dplyr::filter(Scenario == "Historical Reference") %>%
     dplyr::filter(stats::complete.cases(.)) %>%
     dplyr::filter(year == max(year)) %>%
@@ -44,9 +43,11 @@ get_pop_ctry <- function(ssp) {
 
   ssp.data <- rhap::raw.ssp.data %>%
     tidyr::gather(year, value, -Model, -Scenario, -Region, -Variable, -Unit) %>%
-    dplyr::mutate(year=gsub("X", "", year)) %>%
-    dplyr::filter(year >= 2010, year <= 2100,
-                  grepl(ssp, Scenario)) %>%
+    dplyr::mutate(year = gsub("X", "", year)) %>%
+    dplyr::filter(
+      year >= 2010, year <= 2100,
+      grepl(ssp, Scenario)
+    ) %>%
     dplyr::rename(model = Model, scenario = Scenario, region = Region, variable = Variable, unit = Unit) %>%
     dplyr::filter(year >= max_base_year)
 
@@ -57,6 +58,4 @@ get_pop_ctry <- function(ssp) {
     dplyr::select(scenario, region, year, pop)
 
   invisible(pop)
-
 }
-
