+++ tests/testthat/test1-run.R
@@ -7,10 +7,12 @@ test_that("Download db, create project, and run", {
   prj_name <- file.path(rprojroot::find_root(rprojroot::is_testthat), "testInputs", "test_prj_v7p1.dat")
   scen_name <- "Reference"
 
-  testOutput <- calc_hap_impacts(prj_name = prj_name, scen_name = scen_name,
-                                 final_db_year = 2050, HIA_var = "deaths",
-                                 saveOutput = T, map = T, anim = F,
-                                 normalized = F, by_gr = F)
+  testOutput <- calc_hap_impacts(
+    prj_name = prj_name, scen_name = scen_name,
+    final_db_year = 2050, HIA_var = "deaths",
+    saveOutput = TRUE, map = TRUE, anim = FALSE,
+    normalized = FALSE, by_gr = FALSE
+  )
   testResult <- get(load(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/calc_hap_impacts_output1.RData")))
   testthat::expect_equal(nrow(testOutput), nrow(testResult))
 
@@ -18,23 +20,23 @@ test_that("Download db, create project, and run", {
 
   # AUX functions to deal with ASCII characters
   decode_ascii <- function(text) {
-
     escaped_text <- htmltools::htmlEscape(text)
     xml2::xml_text(xml2::read_xml(paste0("<x>", text, "</x>")))
-
   }
 
   testOutput <- read.csv(file.path(rprojroot::find_root(rprojroot::is_testthat), "output/Reference_HAP_deaths.csv"),
-                         fileEncoding = "UTF-8", stringsAsFactors = FALSE)
+    fileEncoding = "UTF-8", stringsAsFactors = FALSEALSE
+  )
   testResult <- read.csv(file.path(rprojroot::find_root(rprojroot::is_testthat), "testOutputs/Reference_HAP_deaths.csv"),
-                         fileEncoding = "UTF-8", stringsAsFactors = FALSE)
+    fileEncoding = "UTF-8", stringsAsFactors = FALSEALSE
+  )
   testOutput$country <- sapply(testOutput$country, decode_ascii)
   testResult$country <- sapply(testResult$country, decode_ascii)
-  testthat::expect(!is.null(testOutput), 'Empty file. Check that the results were correctly produced.')
-  #res <- setdiff(as.data.frame.table(testOutput), as.data.frame.table(testResult))
-  #testthat::expect(nrow(res) == 0, 'Saved results differ from the expected ones. Check that the results were correctly produced and saved.')
-  #res <- setdiff(as.data.frame.table(testResult), as.data.frame.table(testOutput))
-  #testthat::expect(nrow(res) == 0, 'Saved results differ from the expected ones. Check that the results were correctly produced and saved.')
+  testthat::expect(!is.null(testOutput), "Empty file. Check that the results were correctly produced.")
+  # res <- setdiff(as.data.frame.table(testOutput), as.data.frame.table(testResult))
+  # testthat::expect(nrow(res) == 0, 'Saved results differ from the expected ones. Check that the results were correctly produced and saved.')
+  # res <- setdiff(as.data.frame.table(testResult), as.data.frame.table(testOutput))
+  # testthat::expect(nrow(res) == 0, 'Saved results differ from the expected ones. Check that the results were correctly produced and saved.')
 
   # # check figures
   # to be run locally, avoid running it in github actions
@@ -53,53 +55,66 @@ test_that("Download db, create project, and run", {
 
   # error messages
   expect_error(
-    calc_hap_impacts(prj_name = prj_name, scen_name = scen_name,
[…]
