+++ R/calc_hap_impacts.R
@@ -21,12 +21,11 @@
 #' @export
 
 calc_hap_impacts <- function(db_path = NULL, query_path = "./inst/extdata", db_name = NULL, prj_name,
-                            scen_name, queries = "queries_rhap.xml", final_db_year = 2100,
-                            saveOutput = TRUE, map = FALSE, anim = TRUE, HIA_var = "deaths",
-                            normalized = FALSE, by_gr = FALSE) {
-
+                             scen_name, queries = "queries_rhap.xml", final_db_year = 2100,
+                             saveOutput = TRUE, map = FALSE, anim = TRUE, HIA_var = "deaths",
+                             normalized = FALSE, by_gr = FALSE) {
   Country <- country <- sector <- scenario <- region <- year <- group <- ghg <-
-    Units <- value <- adj <-  value_reg <- dec_share <- Pollutant <- `ISO 3` <-
+    Units <- value <- adj <- value_reg <- dec_share <- Pollutant <- `ISO 3` <-
     Percentatge <- iso3 <- iso <- ssp <- `gcam-consumer` <- value_adj <- building <-
     unit <- flsp_m2 <- Model <- Scenario <- Region <- Variable <- Unit <- gdp <-
     value_agg <- share_pop <- gdp_pc <- gdp_agg <- share_gdp <- gdp_dol2011_ppp <-
@@ -85,7 +84,7 @@ calc_hap_impacts <- function(db_path = NULL, query_path = "./inst/extdata", db_n
 
 
   # Ancillary Functions
-  `%!in%` = Negate(`%in%`)
+  `%!in%` <- Negate(`%in%`)
 
   # Add converter: Teragram to kt
   CONV_Tg_kt <- 1E3
@@ -95,33 +94,34 @@ calc_hap_impacts <- function(db_path = NULL, query_path = "./inst/extdata", db_n
 
   # Then, load the rgcam project if prj not passed as a parameter:
   if (!is.null(db_path) & !is.null(db_name)) {
-
-    rlang::inform('Creating project ...')
+    rlang::inform("Creating project ...")
     conn <- rgcam::localDBConn(db_path,
-                               db_name,migabble = FALSE)
+      db_name,
+      migabble = FALSE
+    )
     prj <- suppressWarnings(
       rgcam::addScenario(conn,
-                         prj_name,
-                         scen_name,
-                         paste0(query_path,"/",queries),
-                         saveProj = TRUE)
+        prj_name,
+        scen_name,
+        paste0(query_path, "/", queries),
+        saveProj = TRUE
+      )
     )
-
   } else {
-
-    rlang::inform('Loading project ...')
+    rlang::inform("Loading project ...")
     prj <- rgcam::loadProject(prj_name)
-
   }
   # Consider the final_db_year as the user indicated year or the closes year available in the project file
-  final_db_year<-min(final_db_year,
-                     max(rgcam::getQuery(prj,'nonCO2 emissions by sector (excluding resource production)')$year))
+  final_db_year <- min(
+    final_db_year,
+    max(rgcam::getQuery(prj, "nonCO2 emissions by sector (excluding resource production)")$year)
+  )
 
 
   #-----
   # EXTRACT DATA FROM GCAM SCENARIO OUTPUTS
 
-  rlang::inform('Running rhap ...')
+  rlang::inform("Running rhap ...")
 
   # First, create a database to transform from GCAM_region to country-level:
[…]

[…]
