### 2025-08-18 a981439
capital letters Global South
