### 2025-08-18 836922e
add missing DOI
