+++ paper/paper.bib
@@ -67,7 +67,8 @@
   volume={7},
   number={69},
   pages={3820},
-  year={2022}
+  year={2022},
+  DOI={10.21105/joss.03820}
 }
 
 
