### 2025-08-18 990c42d
add figs internal references
