+++ man/gdp_ctry.SSP4.Rd
@@ -18,8 +18,8 @@ Filtered GDP data per SSP4 (SSP_database_2024.csv).
 }
 \examples{
 \dontrun{
- library(rhap);
- rhap::gdp_ctry.SSP4
+library(rhap)
+rhap::gdp_ctry.SSP4
 }
 }
 \keyword{datasets}
