+++ _pkgdown.yml
@@ -32,6 +32,6 @@ navbar:
      href: https://github.com/bc3LC/rhap
 
 
-      
+
 toc:
   depth: 2
