### 2025-08-20 bc3LC/rhap#2 @jonsampedro
Thank you for your comments @ibarraespinosa!

Given the current progress on the JOSS publication and the version archiving, we will note these suggestions and implement the changes once that process is complete. I hope that works for you!
