### 2025-08-26 #8091 @goldingn
Thanks @klau506!

For issue 3 I only meant to add indentations in snippet of code in the paper. Ie. just enough spaces at the start of [these lines](https://github.com/bc3LC/rhap/blob/main/paper/paper.md?plain=1#L63-L67) that it lines up neatly with the first line. I see you haven't done that yet, since styler didn't touch the paper.md file. You'll probably have to do that manually.

You also didn't make the switch to `remotes` but that's not critical, I don't mind if you prefer `devtools` for some reason!

Re. updating the code. Yes, given you have fixed those Ts and Fs in the codebase with the styler call, I think it's worth making that the version of record. I can see that all the tests are passing so I'm pretty confident styler hasn't broken anything (except for that one thing you fixed!).

You'll need to issue a new GitHub release (e.g. v1.2), update the zenodo archive to point to that release (this may happen automatically but issue a new DOI? I'm not sure), and then let me know on here what the new release version and zenodo link are. Then I can update them here.

Once you've updated that and made those minor paper changes, I'll recommend acceptance and the EiC can do final checks. Nearly there!
