### 2025-08-27 365c9b2
improve paper code snippet
