### 2025-09-18 #8091 @jonsampedro
Dear @goldingn ,

Just a kind reminder regarding my earlier message. We have implemented the suggested changes, created the new release, and shared the Zenodo archive. Please let me know if there’s anything else you’d like us to adjust.

Thank you very much for your time and consideration!

Jon
