+++ paper/paper.bib
@@ -50,7 +50,7 @@
 }
 
 @article{khan2022rmap,
-  title={rmap: An R package to plot and compare tabular data on customizable maps across scenarios and time},
+  title={{rmap}: An R package to plot and compare tabular data on customizable maps across scenarios and time},
   author={Khan, Zarrar and Zhao, Mengqi and Vernon, Chris R and Wild, Thomas and Yarlagadda, Brinda},
   journal={Journal of Open Source Software},
   volume={7},
@@ -61,7 +61,7 @@
 }
 
 @article{sampedro2022rfasst,
-  title={rfasst: An R tool to estimate air pollution impacts on health and agriculture},
+  title={{rfasst}: An R tool to estimate air pollution impacts on health and agriculture},
   author={Sampedro, Jon and Khan, Zarrar and Vernon, Chris R and Smith, Steven J and Waldhoff, Stephanie and Van Dingenen, Rita},
   journal={Journal of Open Source Software},
   volume={7},
