+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{bennitt2025global,
-  title={Global, regional, and national burden of household air pollution, 1990--2021: a systematic analysis for the Global Burden of Disease Study 2021},
+  title={Global, regional, and national burden of household air pollution, 1990--2021: a systematic analysis for the {G}lobal {B}urden of {D}isease Study 2021},
   author={Bennitt, Fiona B and Wozniak, Sarah and Causey, Kate and Spearman, Sandra and Okereke, Chukwuma and Garcia, Vanessa and Hashmeh, Nadim and Ashbaugh, Charlie and Abdelkader, Atef and Abdoun, Meriem and others},
   journal={The Lancet},
   year={2025},
@@ -26,7 +26,7 @@
 }
 
 @article{mohajeri2023urban,
-  title={Urban--rural disparity in global estimation of PM2{\textperiodcentered} 5 household air pollution and its attributable health burden},
+  title={Urban--rural disparity in global estimation of {PM2.5} household air pollution and its attributable health burden},
   author={Mohajeri, Nahid and Hsu, Shih-Che and Milner, James and Taylor, Jonathon and Kiesewetter, Gregor and Gudmundsson, Agust and Kennard, Harry and Hamilton, Ian and Davies, Mike},
   journal={The Lancet Planetary Health},
   volume={7},
@@ -38,7 +38,7 @@
 }
 
 @article{das2021benefits,
-  title={The benefits of action to reduce household air pollution (BAR-HAP) model: A new decision support tool},
+  title={The benefits of action to reduce household air pollution ({BAR-HAP}) model: A new decision support tool},
   author={Das, Ipsita and Lewis, Jessica J and Ludolph, Ramona and Bertram, Melanie and Adair-Rohani, Heather and Jeuland, Marc},
   journal={Plos one},
   volume={16},
@@ -50,7 +50,7 @@
 }
 
 @article{khan2022rmap,
-  title={{rmap}: An R package to plot and compare tabular data on customizable maps across scenarios and time},
+  title={{rmap}: An {R} package to plot and compare tabular data on customizable maps across scenarios and time},
   author={Khan, Zarrar and Zhao, Mengqi and Vernon, Chris R and Wild, Thomas and Yarlagadda, Brinda},
   journal={Journal of Open Source Software},
   volume={7},
@@ -61,7 +61,7 @@
 }
 
 @article{sampedro2022rfasst,
-  title={{rfasst}: An R tool to estimate air pollution impacts on health and agriculture},
+  title={{rfasst}: An {R} tool to estimate air pollution impacts on health and agriculture},
   author={Sampedro, Jon and Khan, Zarrar and Vernon, Chris R and Smith, Steven J and Waldhoff, Stephanie and Van Dingenen, Rita},
   journal={Journal of Open Source Software},
   volume={7},
