### 2025-09-26 #8091 @jonsampedro
Thanks @kthyng !

Yes, the final release [rhap-v1.2](https://github.com/bc3LC/rhap/releases/tag/rhap-v1.2) includes all the changes added during the JOSS review.

I have also corrected the capitalization in my references: [here](https://github.com/bc3LC/rhap/commit/7b63d25a2ca2c1dc6f600220930531605a4aeb05) and [here](https://github.com/bc3LC/rhap/commit/1da46dca1f60b45689865f20940b47602837eba6)

Do you want me to make a new release with these changes? I have not done it yet because v1.2 has already been set as the archive, but whatever is best.
